requirements.txt
# Core Dependencies - LangGraph for orchestration
langgraph==0.2.45
langchain==0.3.7
langchain-community==0.3.5
langchain-huggingface==0.1.2

# LLM & Embeddings (Simplified)
huggingface-hub==0.26.2
sentence-transformers==3.3.1

# Google Gemini (Primary LLM - Fast & High Quality)
google-generativeai>=0.3.0

# Vector Store (FAISS only - removed chromadb)
faiss-cpu==1.13.2

# Search API (Exa only - removed tavily)
exa-py==1.1.3

# Sentiment Analysis (VADER + TextBlob only)
textblob==0.18.0.post0
vaderSentiment==3.3.2
numpy

# Dashboard
streamlit==1.40.2
matplotlib==3.9.3
plotly==5.24.1

# API Server
flask==3.0.0
flask-cors==4.0.0
Flask-APScheduler==1.12.4
Werkzeug>=3.0.0

# Utilities
python-dotenv==1.0.1
requests==2.32.3
pytz==2024.2

# NEW DEPENDENCIES for Upgrade
praw==7.7.1
google-api-python-client==2.111.0
feedparser==6.0.10
redis==5.0.1
pytesseract==0.3.10
Pillow>=10.2.0
statsmodels==0.14.1
prophet==1.1.5
pyjwt==2.8.0
bcrypt==4.1.2
databases==0.9.0
aiosqlite==0.19.0
pydantic>=2.7.4
kaleido==0.2.1
cryptography
apscheduler
transformers
torch
pyjwt


# Enterprise Features
APScheduler==3.10.4
cryptography>=41.0.0
transformers>=4.36.0
torch>=2.1.0

